Moved kitchen to it's own class
