+++ src/me/mrCookieSlime/ExoticGarden/CustomFood.java
@@ -1,6 +1,5 @@
 package me.mrCookieSlime.ExoticGarden;
 
-import me.mrCookieSlime.Slimefun.Lists.RecipeType;
 import me.mrCookieSlime.Slimefun.Objects.Category;
 
 import org.bukkit.entity.Player;
@@ -11,7 +10,7 @@ public class CustomFood extends EGPlant {
 	float food;
 
 	public CustomFood(Category category, ItemStack item, String name, ItemStack[] recipe, int food) {
-		super(category, item, name, RecipeType.KITCHEN, true, recipe);
+		super(category, item, name, Kitchen.EG_KITCHEN, true, recipe);
 		this.food = food;
 	}
